2023/11/13
instead of using a graph traversal algorithm, i switched to representing the bipartite graph as a matrix, and then i do row and column swap operations maintaining two conditions, and seeking a third condition.
this approach will not find the correct final sequence if there are transposable elements, however this is a read quality problem.