2024/01/09
I need to checkpoint a big change. I had to generalise the transposition square because I assumed that when there is a conflict, both entries will have two entries on their rows. however, this isn't always the case, and when it isn't, you need to find make a bunch of cascading swaps. this does eliminate a lot of code that was working.