2024/01/16
funny story: i was in the middle of debugging, ad a bright idea and changed LITERALLY EVERYTHING and it didn't work. upsetting, of course. ctrl + z to go back to the one that seemed to be work, finger accidentally moves off of ctrl and i write z. i'm now stuck. if my working change is ahead of this point, it's gone and i'm not sure what worked and why. anyways i found the working (?) version so checkpointing here. too afraid to see if it actually worked so that's tomorrow's (that's in 10 minutes) problem,