2024/01/26
i have begun testing. i have several output files timing suffix array/sequitur vs dbg/euler. i have started running some tests on the cluster. i also gathered some more data for testing. this is not read data, but rather scaffolds. i removed the spaces and am running it. i'm not hopeful it'll work but might as well try. part of why i doubt it will work is that the scaffolds were of many chromosomes which means there are several start and end points which is not something i accounted for. it does seem like that could be added, but I will not be doing so thank you.

ps: congrats to Izzy and Bleem